Breaking shit
Working on improving the planner view to make it match my vision. I broke the way the tasks are added. I probably have to rework things and not use an offset the way I thought I would. I need to find a way to map the taskcards with time. Wherein the origin of the task card, aka the top border, is on the determined start time, and the bottom of the border is touching the start time + duration.